requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==11.2.1
PyGithub==2.1.1
python-dotenv==1.0.1
django-allauth==0.61.1
django-sitemap-generate==0.1.0
django-ckeditor==6.1.0 
gitpython
gunicorn
markdown
psycopg2
psycopg2-binary
dj-database-url
cloudinary
django-cloudinary-storage
beautifulsoup4
